refactor: JdbcTemplate을 하나로 통합
