+++ src/main/java/next/dao/JdbcTemplate.java
@@ -1,13 +1,15 @@
 package next.dao;
 
 import core.jdbc.ConnectionManager;
-import next.model.User;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 
-public abstract class JdbcTemplate {
+public abstract class JdbcTemplate<T> {
 
     public void update(String sql) {
         Connection con = null;
@@ -38,5 +40,98 @@ public abstract class JdbcTemplate {
         }
     }
 
+    public List<T> query(String sql) {
+        Connection con = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try {
+            con = ConnectionManager.getConnection();
+            pstmt = con.prepareStatement(sql);
+            rs = pstmt.executeQuery();
+
+            List<T> list = new ArrayList<>();
+            try {
+                while(rs.next()) {
+                    list.add(mapRow(rs));
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException("fail to convert resultSet to dao list");
+            }
+            return list;
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } finally {
+            if (rs != null) {
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close resultset");
+                }
+            }
+            if (pstmt != null) {
+                try {
+                    pstmt.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close preparedstatement");
+                }
+            }
+            if (con != null) {
+                try {
+                    con.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close connection");
+                }
+            }
+        }
+        throw new RuntimeException("fail to execute select all query");
+    }
+
+    public T queryForObject(String sql) {
+        Connection con = null;
+        PreparedStatement pstmt = null;
+        ResultSet rs = null;
+        try {
+            con = ConnectionManager.getConnection();
+            pstmt = con.prepareStatement(sql);
+            setValues(pstmt);
+            rs = pstmt.executeQuery();
+
+            try {
+                if (rs.next()) {
+                    return mapRow(rs);
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException("fail to convert resultset to dao");
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } finally {
+            if (rs != null) {
+                try {
+                    rs.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close resultset");
+                }
+            }
+            if (pstmt != null) {
+                try {
+                    pstmt.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close preparedstatement");
+                }
+            }
+            if (con != null) {
+                try {
+                    con.close();
+                } catch (SQLException e) {
+                    throw new RuntimeException("fail to close connection");
+                }
+            }
+        }
+        throw new RuntimeException("fail to execute select query");
+    }
+
     abstract void setValues(PreparedStatement pstmt);
+
+    abstract T mapRow(ResultSet resultSet);
 }
